backend/requirements.txt
aiofiles==23.1.0
asgiref==3.7.2
coloredlogs==15.0.1
Django==4.2.4
djangorestframework==3.14.0
httptools==0.6.0
humanfriendly==10.0
multidict==5.2.0
packaging==23.1
pluggy==1.2.0
prompt-toolkit==3.0.28
pytz==2023.3
rasa-sdk==3.6.1
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
sanic==21.12.2
Sanic-Cors==2.2.0
sanic-routing==0.7.2
sqlparse==0.4.4
typing_extensions==4.7.1
ujson==5.8.0
uvloop==0.17.0
wcwidth==0.2.6
websockets==10.4
